Moved shader object header code
