+++ src/shader-object.cpp
@@ -1,0 +1,8 @@
+#include "shader-object.h"
+
+#include "rhi-shared.h"
+#include "device.h"
+
+namespace rhi {
+
+}
